+++ src/main/java/com/atlassian/db/replica/api/DualConnection.java
@@ -3,7 +3,12 @@ package com.atlassian.db.replica.api;
 import com.atlassian.db.replica.api.reason.Reason;
 import com.atlassian.db.replica.api.state.NoOpStateListener;
 import com.atlassian.db.replica.internal.*;
-import com.atlassian.db.replica.internal.circuitbreaker.BreakOnNotSupportedOperations;
+import com.atlassian.db.replica.internal.ForwardCall;
+import com.atlassian.db.replica.internal.ReadReplicaUnsupportedOperationException;
+import com.atlassian.db.replica.internal.ReplicaCallableStatement;
+import com.atlassian.db.replica.internal.ReplicaConnectionProvider;
+import com.atlassian.db.replica.internal.ReplicaPreparedStatement;
+import com.atlassian.db.replica.internal.ReplicaStatement;
 import com.atlassian.db.replica.internal.circuitbreaker.BreakerConnection;
 import com.atlassian.db.replica.spi.ConnectionProvider;
 import com.atlassian.db.replica.spi.DatabaseCall;
@@ -481,7 +486,7 @@ public final class DualConnection implements Connection {
         private final ConnectionProvider connectionProvider;
         private final ReplicaConsistency consistency;
         private DatabaseCall databaseCall = new ForwardCall();
-        private CircuitBreaker circuitBreaker = new BreakOnNotSupportedOperations();
+        private CircuitBreaker circuitBreaker = new MissingCoverageBreaker(MissingCoverageBreaker.cacheStatically());
         private StateListener stateListener = new NoOpStateListener();
         private Set<String> readOnlyFunctions = new HashSet<>();